refactor: add layout context to store layout state and cache
